Added backend for analyses
Due to the MiniOmni backend not beeing fully open source, we can't build
it our selfs. Therefore we need our own backend, that can receve
Scope-data from the MiniOmni Backend, process it and then forward the
data to the frontend application.
Our Backend is responsible for performing analyses on the data, which
would be to inefficient in the frontend.

Changes:
- Added analysen/README.txt
  => explanation on how to build the backend
- Added backend.c
  => the backend it self
- Added backend.exe
  => precompiled backend for win
- modified .gitignore
  => /analysen/.vscode shouldn't be synced

In the current state the backend can:
- Answer GET /UUID
- Establish a WS connection, but doesn't send data jet